+++ framework/db-core/src/main/java/is/codion/framework/db/package-info.java
@@ -21,6 +21,7 @@
  * <ul>
  * <li>{@link is.codion.framework.db.EntityConnectionProvider#CLIENT_DOMAIN_TYPE}
  * <li>{@link is.codion.framework.db.EntityConnectionProvider#CLIENT_CONNECTION_TYPE}
+ * <li>{@link is.codion.framework.db.EntityConnectionProvider#DESCRIPTION}
  * </ul>
  */
 @org.jspecify.annotations.NullMarked